chore!: remove /bff/v1/journey/single-trip endpoint
